+++ src/javascript/main.js
@@ -17,6 +17,7 @@ require('../components/gallery/script');
 require('../components/html/script');
 require('../components/jecoute/script');
 require('../components/news-push/script');
+require('../components/opening-hours/script');
 require('../components/video/script');
 
 // load our layouts