+++ src/javascript/config.js
@@ -41,5 +41,11 @@ Paris.config = {
   },
   "captcha": {
     "key": "6Lf2DQoTAAAAAKmk3wEFCZuK3FqG00SlQM3o6Yvp"
+  },
+  "mapbox": {
+    "accessToken": "pk.eyJ1IjoicGFyaXNudW1lcmlxdWUiLCJhIjoiY2loZG1vMnYyMDAzNnY0a3FvNG1nNG55biJ9.MP1qcHkEecFGqSTs9gg7cw",
+    "defaultCenter": {"lat": 48.856578, "lon": 2.351828},
+    "defaultBounds": [[48.902157, 2.46976], [48.815575, 2.224122]],
+    "styleLayer": "mapbox://styles/parisnumerique/cis1rkqoj000khinpthppoaqd"
   }
 };