+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Let's take the `x.β.α` version number:
 
 ## x.β.α (coming soon)
 
+* component `expandable` added
 * component `opening-hours`: changed data structure
 * component `rush-hours`: changed data structure
 * re-added default font size for body