+++ CHANGELOG.md
@@ -9,6 +9,16 @@ Let's take the `x.β.α` version number:
 
 ## x.β.α (coming soon)
 
+* component `card`: changed data structure (removed `src` in `img`)
+* component `events` added
+* component `expandable` added
+* component `map` added
+* component `opening-hours`: changed data structure
+* component `rush-hours`: changed data structure
+* component `text`: fixed font size in lists (same as paragraphs)
+* module `block-aside-access` removed
+* re-added default font size for body
+
 ## 1.9.4 (2017-10-10)
 
 * module `block-content-newsletter`: add a link to mentions légales