+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Let's take the `x.β.α` version number:
 
 ## x.β.α (coming soon)
 
+## 1.9.5 (2017-11-13)
+
 * component `card`: changed data structure (removed `src` in `img`)
 * component `events` added
 * component `expandable` added